[libc] move abs_timesout and monotonicity out of linux dir
